+++ src/views/Catalog.vue
@@ -256,14 +256,14 @@ export default defineComponent({
         });
       }
     },
-    zoomToBbox({ bbox, crs }) {
+    zoomToBbox({ bbox, crs, crsDefinition }) {
       if (this.$refs.mapView?.zoomToBbox) {
-        this.$refs.mapView.zoomToBbox(bbox, crs);
+        this.$refs.mapView.zoomToBbox(bbox, crs, crsDefinition);
       }
     },
-    highlightBbox({ bbox, crs }) {
+    highlightBbox({ bbox, crs, crsDefinition }) {
       if (this.$refs.mapView?.highlightBbox) {
-        this.$refs.mapView.highlightBbox(bbox, crs);
+        this.$refs.mapView.highlightBbox(bbox, crs, crsDefinition);
       }
     },
     async filterItems(filters, reset) {